backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-sslserver==0.22
cryptography==41.0.7
python-magic==0.4.27
Pillow==10.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
pyotp==2.9.0
qrcode==7.4.2 